requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
litemapy>=0.9.0
requests>=2.25.0
